:recycle: 타입 분리 및 버그 해결
