+++ frontend/src/components/Course/CourseList/CourseList.tsx
@@ -1,6 +1,6 @@
 import { FaWalking } from '@react-icons/all-files/fa/FaWalking';
+import { Course, CourseDetails } from '@/pages/Course/Course.type';
 import CourseItem from '@/components/Course/CourseItem/CourseItem';
-import { CourseDetails } from '@/apis/course';
 import { COURSE_TEXT } from '@/constants/course';
 import * as S from './CourseList.css';
 
@@ -11,7 +11,7 @@ interface CourseListProps {
 const CourseList = ({ courseList }: CourseListProps) => {
   return (
     <>
-      {Object.entries(courseList).map(([key, value]) => (
+      {Object.entries(courseList).map(([key, value]: [string, Course]) => (
         <div key={key}>
           {value !== null && (
             <>
@@ -31,7 +31,7 @@ const CourseList = ({ courseList }: CourseListProps) => {
                       <FaWalking size="15px" />
                     </div>
                   </div>
-                  걸어서 10분
+                  걸어서 {value.pedestrian_route.duration_minutes}분
                 </div>
               )}
             </>
