+++ frontend/src/pages/Course/Course.type.ts
@@ -1,3 +1,10 @@
+export interface CourseDetails {
+  lunch: Course | null;
+  cafe: Course | null;
+  attraction: Course | null;
+  dinner: Course | null;
+}
+
 export interface Course {
   id: string;
   phone: string;
@@ -5,4 +12,7 @@ export interface Course {
   road_address_name: string;
   x: string;
   y: string;
+  pedestrian_route: {
+    duration_minutes: number;
+  };
 }
